Split HubRestApi mega-class into multiple files
