+++ .vscode/settings.json
@@ -1,0 +1,3 @@
+{
+    "python.pythonPath": "c:\\Users\\arcalder\\Projects\\hub-rest-api-python\\virtualenv\\Scripts\\python.exe"
+}
