+++ blackduck/Authentication.py
@@ -1,0 +1,7 @@
+import logging
+import requests
+import json
+from operator import itemgetter
+import urllib.parse
+
+logger = logging.getLogger(__name__)
