+++ blackduck/__init__.py
@@ -1,1 +1,2 @@
 
+from .HubRestApi import HubInstance
